requirements.txt
flask
flask-restful
Flask-Migrate
flask-cors
Flask-Script
Flask-SQLAlchemy
python-dotenv